dont mix up device_chunk_size and chunk_size

+++ src/multiexp.rs
@@ -348,7 +348,7 @@ fn test_with_bls12() {
 
 pub fn gpu_multiexp_supported<E>() -> gpu::GPUResult<gpu::MultiexpKernel<E>> where E: Engine {
     const TEST_SIZE : u32 = 1024;
-    const CHUNK_SIZE: u32 = 8388608;
+    const CHUNK_SIZE: usize = 8388608;
     use rand::Rand;
     let pool = Worker::new();
     let rng = &mut rand::thread_rng();
@@ -374,7 +374,7 @@ pub fn gpu_multiexp_consistency() {
     const CHUNK_SIZE: usize = 1048576;
     const MAX_LOG_D: usize = 20;
     const START_LOG_D: usize = 10;
-    let mut kern = gpu::MultiexpKernel::<Bls12>::create(CHUNK_SIZE as u32).ok();
+    let mut kern = gpu::MultiexpKernel::<Bls12>::create(CHUNK_SIZE).ok();
     if kern.is_none() { panic!("Cannot initialize kernel!"); }
     let pool = Worker::new();
